bugfix: make receiving state more robust
The 'receiving' state is a bit special, because it can be triggered
indepentent of the current state (by incoming data). After the state is
handled, the state machine should therefore return to its previous
state.

However, the 'receiving' state also calls a user callback which
may also change the state, causing a conflict.

This bugfix makes sure the state machine only returns to a previous
state if it detects that the user callback has not also changed it.